Centralize Fathom media write fence
