+++ packages/twenty-apps/public/fathom/src/logic-functions/utils/build-fathom-media-write-fence.util.ts
@@ -1,0 +1,13 @@
+import { type FathomMediaWriteContext } from 'src/logic-functions/types/fathom-media-write-context.type';
+import { isDefined } from 'src/utils/is-defined';
+
+export const buildFathomMediaWriteFence = (
+  writeContext: FathomMediaWriteContext,
+) => ({
+  id: { eq: writeContext.fathomRecordingImportId },
+  connectedAccountId: { eq: writeContext.connectedAccountId },
+  mediaImportClaimedAt: { eq: writeContext.claimedAt },
+  mediaDownloadId: isDefined(writeContext.downloadId)
+    ? { eq: writeContext.downloadId }
+    : { is: 'NULL' as const },
+});
